requirementes.txt
djangorestframework
djangorestframework-simplejwt
itsdangerous
